api/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-core>=0.1.0
psycopg2-binary>=2.9.9
pgvector>=0.2.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0
sentence-transformers>=2.2.0
langchain-huggingface>=0.0.1

